// api/src/extraction/strategies/grill-area/grill-area.interface.ts
import { ExtractionResult } from '../strategy.interface';

export const GrillAreaSourceType = 'grill-area';

export type GrillAreaConfig = {
  from: Date;
  to: Date;
  areas: number[];
  weekdays?: string[];
};

export type GrillAreaSource = {
  type: typeof GrillAreaSourceType;
  config: GrillAreaConfig;
};

export type GrillAreaResult = ExtractionResult & {
  area: number;
  day: Date;
};